Split fixtures in the Thumbnail module
