+++ conftest.py
@@ -38,6 +38,7 @@ pytest_plugins = [
     "saleor.page.tests.fixtures",
     "saleor.menu.tests.fixtures",
     "saleor.warehouse.tests.fixtures",
+    "saleor.thumbnail.tests.fixtures",
 ]
